+++ CHANGELOG.md
@@ -7,6 +7,7 @@ principles.
 ## unreleased
 - Fixed: issue with sidebar trigger when using nested forms
 - Added: option to choose where to go to after creating model. Via `Resource::getRedirectAfterCreate()`. This can be set to null as well aka when used in sidebar.
+- Added: resource::getInstanceAttributes method to set default attributes on a model when record is not created yet
 
 ## 0.7.7 - 2022-05-17
 - Fixed: Trigger PageChanged event when fragments are sorted.